Use 64-bit duty math and fix header newline

+++ src/ESC.cpp
@@ -17,6 +17,6 @@ void ESC::arm(int pulse) {
 void ESC::writeMicroseconds(int pulse) {
     pulse = constrain(pulse, 1000, 2000);
     uint32_t maxDuty = (1 << _resolution) - 1;
-    uint32_t duty = (uint64_t)maxDuty * pulse * _freq / 1000000; // scale to period
-    ledcWrite(_channel, duty);
+    uint64_t duty = (uint64_t)maxDuty * pulse * _freq / 1000000; // scale to period
+    ledcWrite(_channel, (uint32_t)duty);
 }
